BugImpact added

+++ src/Model/BugReport/BugReport.java
@@ -293,7 +293,7 @@ public class BugReport extends Subject implements Observer, Originator<BugReport
         return this.impactFactor;
     }
 
-    public int getBugImpact() {
+    public double getBugImpact() {
         return this.getTag().getBugImpact(this);
     }
